+++ game_patch/main/main.cpp
@@ -364,7 +364,7 @@ void load_config()
     // Log information from config
     xlog::info("Resolution: {}x{}x{}", g_game_config.res_width.value(), g_game_config.res_height.value(), g_game_config.res_bpp.value());
     xlog::info("Window Mode: {}", static_cast<int>(g_game_config.wnd_mode.value()));
-    xlog::info("Max FPS: {}", g_game_config.max_fps.value());
+    //xlog::info("Max FPS: {}", g_game_config.max_fps.value());
     xlog::info("Allow Overwriting Game Files: {}", g_game_config.allow_overwrite_game_files.value());
 }